Fix ts issue with FormOptions type name
