+++ packages/bodiless-core/src/components/ContextMenuItem.tsx
@@ -12,51 +12,19 @@
  * limitations under the License.
  */
 
-import React, {
-  useState, createContext, useContext,
-} from 'react';
-import ReactTooltip from 'rc-tooltip';
-import { getUI as getFormUI } from '../contextMenuForm';
-import { useContextMenuContext } from './ContextMenuContext';
-import type { UI, IContextMenuItemProps as IProps, ContextMenuFormProps } from '../Types/ContextMenuTypes';
-
-const defaultUI = {
-  Icon: 'span',
-  ToolbarButton: 'div',
-  FormWrapper: 'div',
-  ToolbarDivider: 'div',
-  Form: 'form',
-  Tooltip: ReactTooltip,
-};
-
-export const getUI = (ui: UI = {}) => ({
-  ...defaultUI,
-  ...getFormUI(),
-  ...ui,
-});
-
-const UIContext = createContext<UI>({});
-export const useUI = () => {
-  const ui = useContext(UIContext);
-  return getUI(ui);
-};
+import React, { useState } from 'react';
+import { useContextMenuContext, useContextMenuUIContext } from './ContextMenuContext';
+import type { IContextMenuItemProps as IProps, ContextMenuFormProps } from '../Types/ContextMenuTypes';
 
 const ContextMenuItem = (props: IProps) => {
-  const {
-    option,
-    index,
-    ui,
-  } = props;
+  const { option, index } = props;
   const [renderForm, setRenderForm$] = useState<(props:ContextMenuFormProps) => JSX.Element>();
   const [isToolTipShown, setIsToolTipShown] = useState(false);
-  const finalUI = getUI(ui);
+  const ui = useContextMenuUIContext();
   const {
-    ToolbarDivider,
-    Icon,
-    ToolbarButton,
-    FormWrapper,
-    Tooltip,
-  } = finalUI;
+    ToolbarDivider, Icon, ToolbarButton,
+    FormWrapper, Tooltip,
+  } = ui;
   const isActive = option.isActive ? option.isActive() : false;
   const isDisabled = option.isDisabled ? option.isDisabled() : false;
   const isHidden = option.isHidden ? option.isHidden() : false;
@@ -85,14 +53,12 @@ const ContextMenuItem = (props: IProps) => {
     if (renderForm) {
       const formProps: ContextMenuFormProps = {
         closeForm: onFormClose,
-        ui: finalUI,
+        ui,
         'aria-label': `Context Menu ${option.label || option.name} Form`,
       };
       return (
         <FormWrapper onClick={(e: React.MouseEvent<HTMLDivElement>) => e.stopPropagation()}>
-          <UIContext.Provider value={finalUI}>
-            {renderForm(formProps)}
-          </UIContext.Provider>
+          {renderForm(formProps)}
         </FormWrapper>
       );
     }
