+++ packages/bodiless-core/src/components/index.ts
@@ -12,12 +12,13 @@
  * limitations under the License.
  */
 
-import ContextMenu, { getUI } from './ContextMenu';
+import ContextMenu from './ContextMenu';
 import ContextWrapper from './ContextWrapper';
 import LocalContextMenu from './LocalContextMenu';
 import PageEditor from './PageEditor';
 import StaticPage from './StaticPage';
 import PageOverlay from './PageOverlay';
+import { getUI } from './ContextMenuContext';
 import type {
   IContextMenuProps as ContextMenuProps,
   ButtonVariantProps,
