+++ packages/bodiless-core/src/index.ts
@@ -61,7 +61,7 @@ import {
 import withChild from './withChild';
 import asBodilessComponent, { withActivatorWrapper } from './asBodilessComponent';
 import type { Options as BodilessOptions, AsBodiless } from './asBodilessComponent';
-import { useUI as useFormUI } from './components/ContextMenuItem';
+import { useContextMenuUIContext as useFormUI } from './components/ContextMenuContext';
 import ContextSubMenu from './ContextMenu/ContextSubMenu';
 
 export * from './components';
