+++ packages/bodiless-core/src/withNotificationButton.tsx
@@ -16,11 +16,11 @@ import React, { useCallback } from 'react';
 import { ContextMenuForm } from './contextMenuForm';
 import { useRegisterMenuOptions } from './PageContextProvider';
 import { useNotifications } from './NotificationProvider';
-import { useUI as useFormUI } from './components/ContextMenuItem';
+import { useContextMenuUIContext } from './components/ContextMenuContext';
 import type { ContextMenuFormProps } from './Types/ContextMenuTypes';
 
 const NotificationList = () => {
-  const { ComponentFormList, ComponentFormListItem } = useFormUI();
+  const { ComponentFormList, ComponentFormListItem } = useContextMenuUIContext();
   const { notifications } = useNotifications();
   if (notifications.length === 0) return (<p>There are no alerts.</p>);
   return (
@@ -35,7 +35,7 @@ const NotificationList = () => {
 };
 
 const RenderForm = (props: ContextMenuFormProps) => {
-  const { ComponentFormTitle } = useFormUI();
+  const { ComponentFormTitle } = useContextMenuUIContext();
   return (
     <ContextMenuForm {...props}>
       <ComponentFormTitle>Alerts</ComponentFormTitle>
